Initial import of karvig library.
Import karvig library that will be used as a personal playground as well
as a possibility of this library to become a personal toolchain/kit.
Currently implemented features include (actually just one for now) - is
a naive partial Singly Liked List with support for generic datatypes.
Lots of code clean up as well as documentation and a probable
re-implementation remain.

Other utils that I plan to put in here include:
    1. Doubly Linked List.
    2. Queues.
    3. Stack.
    4. Dictionary.
    5. String utils.