Add node-free callback.
Add support for a callback for free-ing user defined data. This is
useful when the user of the library wants more fine grained control of
what is done to the data (rather than calling a plain free()). Refer to
documentation of slist_set_node_free_func() for the contract defined for
using this callback.